refactor: 도서 물리 삭제 고도화

+++ src/main/java/com/part3/team07/sb01deokhugamteam07/service/BookService.java
@@ -71,8 +71,10 @@ public class BookService {
   }
 
   public void hardDelete(UUID id) {
-    bookRepository.existsById(id);
-
+    if (!bookRepository.existsById(id)) {
+      throw new BookNotFoundException();
+    }
+    
     bookRepository.deleteById(id);
   }
 }
